+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+- Fix #62: ensure that a child component's `mount()` lifecycle method is only called after its parent has completely finished rendering
+
 # 3.1.1
 - Fix: components that changed their root HTML tag were erroneously unmounted